requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow==2.20.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.4
werkzeug==3.0.1
python-dateutil==2.8.2
google-generativeai==0.3.1
